requirements.txt
seaborn==0.12.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
telethon==1.36.0
psycopg2-binary==2.9.9
dbt-core==1.8.0
dbt-postgres==1.8.0
ultralytics==8.2.0
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.7.0
python-dotenv==1.0.1
dagster==1.7.0
dagster-webserver==1.7.0
pytest==8.2.0
scikit-learn==1.4.2
sqlalchemy==2.0.30
# python:3.10-slim